test(workflow): localize lifecycle transition coverage
